Important commit with two major changes:
(1) The way XML is treated is more consistent, with one class for every node type in the xml. This means that the XML writing part is decentralized and we have a lot of classes that are used to store data naturally.
 The sbml and uniprot parser have been rewritten to use theses new classes, all the older classes used for storage have been removed.

(2) New RBA output files are being produced. The metabolism file has its own format (it does not depend on SBML anymore). A standard parameter file is generated and a half of the process file is also being generated.